added more realistic bullets

+++ include/gameHandler.h
@@ -2,7 +2,7 @@
 #include "datatypes.h"
 #include <SDL2/SDL.h>
 
-void createGameHandler(GameHandler* gameHandler, Helecopter* helecopter, Vec2 helecopterScreenCoords, int groundHight, double gameScale);
+void createGameHandler(GameHandler* gameHandler, Helecopter* helecopter, Vec2 helecopterScreenCoords, double groundHight, double gameScale);
 void gameHandlerUpdate(GameHandler* gameHandler, Vec2 windowSize);
 void addexplosion(GameHandler* gameHandler, Vec2f coords, double size);
 void destroyExplosion(GameHandler* gameHandler, int element);
